ml_server/requirements.txt
fastapi
uvicorn[standard]
ultralytics==8.3.34
pillow
numpy
opencv-python-headless
